Refactor logging and formatting
