+++ archinstall/applications/fonts.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING
 
+from archinstall.lib.log import debug
 from archinstall.lib.models.application import FontsConfiguration
-from archinstall.lib.output import debug
 
 if TYPE_CHECKING:
 	from archinstall.lib.installer import Installer
