+++ archinstall/lib/configuration.py
@@ -8,10 +8,10 @@ from pydantic import TypeAdapter
 
 from archinstall.lib.args import ArchConfig, ArchConfigType
 from archinstall.lib.crypt import encrypt
+from archinstall.lib.log import debug, logger, warn
 from archinstall.lib.menu.helpers import Confirmation, Selection
 from archinstall.lib.menu.util import get_password, prompt_dir
 from archinstall.lib.models.network import NetworkConfiguration
-from archinstall.lib.output import debug, logger, warn
 from archinstall.lib.translationhandler import tr
 from archinstall.lib.utils.format import as_key_value_pair
 from archinstall.tui.menu_item import MenuItem, MenuItemGroup
