+++ archinstall/lib/disk/fido.py
@@ -4,9 +4,9 @@ from typing import ClassVar
 
 from archinstall.lib.command import SysCommand, SysCommandWorker
 from archinstall.lib.exceptions import SysCallError
+from archinstall.lib.log import error, info
 from archinstall.lib.models.device import Fido2Device
 from archinstall.lib.models.users import Password
-from archinstall.lib.output import error, info
 from archinstall.lib.utils.encoding import clear_vt100_escape_codes_from_str
