+++ archinstall/lib/general/general_menu.py
@@ -1,8 +1,8 @@
 from enum import Enum
 
 from archinstall.lib.locale.utils import list_timezones
+from archinstall.lib.log import warn
 from archinstall.lib.menu.helpers import Confirmation, Input, Selection
-from archinstall.lib.output import warn
 from archinstall.lib.translationhandler import Language, tr
 from archinstall.tui.menu_item import MenuItem, MenuItemGroup
 from archinstall.tui.result import ResultType
