+++ archinstall/lib/menu/abstract_menu.py
@@ -2,8 +2,8 @@ from enum import Enum
 from types import TracebackType
 from typing import Any, Self, override
 
+from archinstall.lib.log import error
 from archinstall.lib.menu.helpers import Selection
-from archinstall.lib.output import error
 from archinstall.lib.translationhandler import tr
 from archinstall.tui.components import InstanceRunnable
 from archinstall.tui.menu_item import MenuItem, MenuItemGroup
