+++ archinstall/lib/mirror/mirror_handler.py
@@ -2,10 +2,10 @@ import time
 import urllib.parse
 from pathlib import Path
 
+from archinstall.lib.log import debug, info
 from archinstall.lib.models import MirrorRegion
 from archinstall.lib.models.mirrors import MirrorStatusEntryV3, MirrorStatusListV3
 from archinstall.lib.networking import fetch_data_from_url
-from archinstall.lib.output import debug, info
 from archinstall.lib.pathnames import MIRRORLIST
