+++ archinstall/lib/models/mirrors.py
@@ -9,10 +9,10 @@ from typing import TYPE_CHECKING, Any, Self, TypedDict, override
 
 from pydantic import BaseModel, ValidationInfo, field_validator, model_validator
 
+from archinstall.lib.log import debug
 from archinstall.lib.models.config import SubConfig
 from archinstall.lib.models.packages import Repository
 from archinstall.lib.networking import DownloadTimer, ping
-from archinstall.lib.output import debug
 from archinstall.lib.translationhandler import tr
 
 if TYPE_CHECKING:
