+++ archinstall/lib/network/wifi_handler.py
@@ -5,9 +5,9 @@ from typing import assert_never, override
 
 from archinstall.lib.command import SysCommand
 from archinstall.lib.exceptions import SysCallError
+from archinstall.lib.log import debug
 from archinstall.lib.models.network import WifiConfiguredNetwork, WifiNetwork
 from archinstall.lib.network.wpa_supplicant import WpaSupplicantConfig
-from archinstall.lib.output import debug
 from archinstall.lib.translationhandler import tr
 from archinstall.tui.components import ConfirmationScreen, InputScreen, InstanceRunnable, LoadingScreen, NotifyScreen, TableSelectionScreen, tui
 from archinstall.tui.menu_item import MenuItem, MenuItemGroup
