+++ archinstall/lib/network/wpa_supplicant.py
@@ -1,8 +1,8 @@
 from dataclasses import dataclass, field
 from pathlib import Path
 
+from archinstall.lib.log import debug
 from archinstall.lib.models.network import WifiNetwork
-from archinstall.lib.output import debug
 
 
 @dataclass
