+++ archinstall/lib/packages/packages.py
@@ -1,9 +1,9 @@
 from functools import lru_cache
 
 from archinstall.lib.exceptions import SysCallError
+from archinstall.lib.log import debug
 from archinstall.lib.menu.helpers import Loading, Notify, Selection
 from archinstall.lib.models.packages import AvailablePackage, LocalPackage, PackageGroup, Repository
-from archinstall.lib.output import debug
 from archinstall.lib.pacman.pacman import Pacman
 from archinstall.lib.translationhandler import tr
 from archinstall.tui.menu_item import MenuItem, MenuItemGroup
