+++ archinstall/scripts/minimal.py
@@ -4,12 +4,12 @@ from archinstall.lib.configuration import ConfigurationOutput
 from archinstall.lib.disk.disk_menu import DiskLayoutConfigurationMenu
 from archinstall.lib.disk.filesystem import FilesystemHandler
 from archinstall.lib.installer import Installer
+from archinstall.lib.log import debug, error, info
 from archinstall.lib.menu.util import delayed_warning
 from archinstall.lib.models import Bootloader
 from archinstall.lib.models.profile import ProfileConfiguration
 from archinstall.lib.models.users import Password, User
 from archinstall.lib.network.network_handler import install_network_config
-from archinstall.lib.output import debug, error, info
 from archinstall.lib.profile.profiles_handler import profile_handler
 from archinstall.lib.translationhandler import tr
 from archinstall.tui.components import tui
